+++ shiny/Data-Ingest/server_files/d1_download_svr.R
@@ -20,6 +20,8 @@ observeEvent(input$D1Button, {
 # Display downloaded files in data.frame
 output$identifier <- DT::renderDT({Shared.data$downloaded})
 
+output$rowSelection <- renderPrint(input$identifier_rows_selected)
+
 # Move files to correct dbfiles location (make a custom function for this?)
 observeEvent(input$D1FinishButton, {
   # create the new directory in /dbfiles
@@ -31,3 +33,5 @@ observeEvent(input$D1FinishButton, {
   }
   output$D1dbfilesPath <- renderText({paste0(PEcAn_path, d1_dirname)}) # Print path to data
 })
+
+