+++ shiny/Data-Ingest/ui_files/sidebar_ui.R
@@ -1,9 +1,9 @@
 sidebarMenu(
-  # menuItem(
-  #   "Select Data Acquisition Method",
-  #   tabName = "importDataONE",
-  #   icon = icon("file")
-  # ),
+  menuItem(
+    "Homepage",
+    tabName = "Home",
+    icon = icon("home", lib = "font-awesome")
+  ),
   menuItem(
     "Import from DataONE",
     tabName = "importDataONE",
@@ -13,19 +13,10 @@ sidebarMenu(
     "Upload Local Files",
     tabName = "uploadLocal",
     icon = icon("upload", lib = "font-awesome")
-  ),
-  menuItem(
-    "Step 2 -- New Input",
-    tabName = "step2",
-    icon = icon("database", lib = "font-awesome")
-  ),
-  menuItem(
-    "Step 3 -- format record",
-    tabName = "step3",
-    icon = icon("cog")
-  ),
-  menuItem("Step 4 -- etc.", 
-           tabName = "step4", 
-           icon = icon("cog")
   )
+  # menuItem(
+  #   "Step 2 -- New Input",
+  #   tabName = "step2",
+  #   icon = icon("database", lib = "font-awesome")
+  # )
 )