+++ CHANGELOG.md
@@ -45,6 +45,7 @@ For more information about this file see also [Keep a Changelog](http://keepacha
   - Users can now input the name of the destination directory that they wish to create within dbfiles. 
   - Updated Travis.yml to include librdf0-dev so that it can download redland, datapack, and dataone. 
   - Added Data-Ingest UI (inputs, dbfiles, and formats record UI and some basic server side functionality are online)
+  - Modularized input record, format record, and dbfiles record into shiny modules. This allows the app to be greatly simplified to two, single-page workflows. These functions can also be used "plug-and-play" style elsewhere in PEcAn shiny apps to load in data. 
 
   
 ### Removed